fix: move to project folder
