+++ packages/project-access/src/index.ts
@@ -19,6 +19,9 @@ export {
     getCdsServices,
     getCapI18nFolderNames,
     getI18nPropertiesPaths,
+    getMinUI5VersionFromManifest,
+    getMinUI5VersionAsArray,
+    getMinimumUI5Version,
     getMtaPath,
     getNodeModulesPath,
     getProject,
