+++ packages/project-access/src/library/index.ts
@@ -1,7 +1,4 @@
 export {
     getReuseLibs,
-    checkDependencies,
-    getMinUI5VersionFromManifest,
-    getMinUI5VersionAsArray,
-    getMinimumUI5Version
+    checkDependencies
 } from './helpers';
