+++ packages/project-access/src/project/index.ts
@@ -15,7 +15,15 @@ export {
 } from './cap';
 export { getNodeModulesPath } from './dependencies';
 export { getCapI18nFolderNames, getI18nPropertiesPaths } from './i18n';
-export { getAppProgrammingLanguage, getAppType, getProject, getProjectType } from './info';
+export {
+    getAppProgrammingLanguage,
+    getAppType,
+    getMinUI5VersionFromManifest,
+    getMinUI5VersionAsArray,
+    getMinimumUI5Version,
+    getProject,
+    getProjectType
+} from './info';
 export { loadModuleFromProject } from './module-loader';
 export { findAllApps, findCapProjects, findFioriArtifacts, findProjectRoot, getAppRootFromWebappPath } from './search';
 export { getWebappPath, readUi5Yaml } from './ui5-config';
