+++ packages/project-access/test/library/helpers.test.ts
@@ -3,10 +3,7 @@ import {
     checkDependencies,
     getReuseLibs,
     getLibraryDesc,
-    getManifestDesc,
-    getMinUI5VersionFromManifest,
-    getMinUI5VersionAsArray,
-    getMinimumUI5Version
+    getManifestDesc
 } from '../../src/library/helpers';
 import * as manifestJson from '../test-data/libs/sap.reuse.ex.test.lib.attachmentservice/src/sap/reuse/ex/test/lib/attachmentservice/manifest.json';
 import type { Manifest, ReuseLib } from '../../src';
@@ -104,91 +101,5 @@ describe('library utils', () => {
         };
         const description = await getLibraryDesc(lib, 'mock/path');
         expect(description).toEqual('test description');
-    });
-
-    describe('getters for minUI5Version', () => {
-        interface TestCase {
-            minUI5Version: string | string[] | undefined;
-            expectedResult: string | string[] | undefined;
-        }
-
-        describe('getMinUI5VersionFromManifest', () => {
-            const testVersions: Array<TestCase> = [
-                { minUI5Version: undefined, expectedResult: undefined },
-                { minUI5Version: 'a.b.c', expectedResult: 'a.b.c' },
-                { minUI5Version: ['a.b.c'], expectedResult: ['a.b.c'] },
-                { minUI5Version: '1.76.32', expectedResult: '1.76.32' },
-                { minUI5Version: '1.76', expectedResult: '1.76' },
-                { minUI5Version: ['1.120.3'], expectedResult: ['1.120.3'] },
-                { minUI5Version: ['1.125'], expectedResult: ['1.125'] },
-                { minUI5Version: ['1.120.13', 'a.b.c.'], expectedResult: ['1.120.13', 'a.b.c.'] }
-            ];
-
-            testVersions.forEach((testCase) => {
-                test(`getMinUI5VersionFromManifest: minUI5Version = ${testCase.minUI5Version}`, () => {
-                    const manifest = {
-                        'sap.ui5': {
-                            dependencies: {
-                                minUI5Version: testCase.minUI5Version
-                            }
-                        }
-                    } as Manifest;
-                    expect(getMinUI5VersionFromManifest(manifest)).toEqual(testCase.expectedResult);
-                });
-            });
-        });
-
-        describe('getMinUI5VersionAsArray', () => {
-            const testVersions: Array<TestCase> = [
-                { minUI5Version: undefined, expectedResult: [] },
-                { minUI5Version: 'a.b.c', expectedResult: ['a.b.c'] },
-                { minUI5Version: ['a.b.c'], expectedResult: ['a.b.c'] },
-                { minUI5Version: '1.76.32', expectedResult: ['1.76.32'] },
-                { minUI5Version: '1.76', expectedResult: ['1.76'] },
-                { minUI5Version: ['1.120.3'], expectedResult: ['1.120.3'] },
-                { minUI5Version: ['1.125'], expectedResult: ['1.125'] },
-                { minUI5Version: ['1.120.13', 'a.b.c.'], expectedResult: ['1.120.13', 'a.b.c.'] }
-            ];
-
-            testVersions.forEach((testCase) => {
-                test(`getMinUI5VersionAsArray: minUI5Version = ${testCase.minUI5Version}`, () => {
-                    const manifest = {
-                        'sap.ui5': {
-                            dependencies: {
-                                minUI5Version: testCase.minUI5Version
-                            }
-                        }
-                    } as Manifest;
-                    expect(getMinUI5VersionAsArray(manifest)).toEqual(testCase.expectedResult);
-                });
-            });
-        });
-
-        describe('getMinimumUI5Version', () => {
-            const testVersions: Array<TestCase> = [
-                { minUI5Version: undefined, expectedResult: undefined },
-                { minUI5Version: 'a.b.c', expectedResult: undefined },
-                { minUI5Version: ['a.b.c'], expectedResult: undefined },
-                { minUI5Version: '1.76.32', expectedResult: '1.76.32' },
-                { minUI5Version: '1.76', expectedResult: undefined },
-                { minUI5Version: ['1.120.3'], expectedResult: '1.120.3' },
-                { minUI5Version: ['1.120.13', 'a.b.c.'], expectedResult: '1.120.13' },
-                { minUI5Version: ['1.120.13', '2.0.21'], expectedResult: '1.120.13' },
-                { minUI5Version: ['2.0.21', '1.120.13'], expectedResult: '1.120.13' }
-            ];
-
-            testVersions.forEach((testCase) => {
-                test(`getMinimumUI5Version: minUI5Version = ${testCase.minUI5Version}`, () => {
-                    const manifest = {
-                        'sap.ui5': {
-                            dependencies: {
-                                minUI5Version: testCase.minUI5Version
-                            }
-                        }
-                    } as Manifest;
-                    expect(getMinimumUI5Version(manifest)).toEqual(testCase.expectedResult);
-                });
-            });
-        });
     });
 });
